Extract model display helpers from Utility into Math::ModelHelper
