+++ lib/plurimath/utility.rb
@@ -115,11 +115,6 @@ module Plurimath
       "{": "}",
       "[": "]",
     }.freeze
-    TEXT_CLASSES = %w[
-      symbol
-      number
-      text
-    ].freeze
     UNICODEMATH_MENCLOSE_FUNCTIONS = {
       underline: "bottom",
       underbar: "bottom",
@@ -336,76 +331,6 @@ module Plurimath
         symbols_class(value, lang: lang)
       end
 
-      def validate_left_right(fields = [])
-        fields.each do |field|
-          if field.is_a?(Math::Formula) && field.value.first.is_a?(Math::Function::Left)
-            field.left_right_wrapper = true
-          end
-        end
-      end
-
-      def validate_math_zone(object, lang:, intent: false, options: nil)
-        return false unless object
-
-        if object.is_a?(Math::Formula)
-          filter_math_zone_values(object.value, lang: lang, intent: intent,
-                                                options: options).find do |value|
-            !(value.is_a?(Math::Function::Text) || value.is_a?(Math::Symbols::Symbol))
-          end
-        else
-          !(TEXT_CLASSES.include?(object.class_name) || object.is_a?(Math::Symbols::Symbol))
-        end
-      end
-
-      def filter_math_zone_values(value, lang:, intent: false, options: nil)
-        return [] if value&.empty?
-
-        new_arr = []
-        temp_array = []
-        skip_index = nil
-        value.each_with_index do |obj, index|
-          object = obj.dup
-          next if index == skip_index
-          if TEXT_CLASSES.include?(object.class_name) || math_display_text_objects(object)
-            next temp_array << (if object.is_a?(Math::Symbols::Symbol)
-                                  symbol_to_text(
-                                    object, lang: lang, intent: intent, options: options
-                                  )
-                                else
-                                  object.value
-                                end)
-          end
-
-          if temp_array.any?
-            new_arr << Math::Function::Text.new(temp_array.join(" "),
-                                                lang: lang)
-          end
-          temp_array = []
-          new_arr << object
-        end
-        if temp_array.any?
-          new_arr << Math::Function::Text.new(temp_array.join(" "),
-                                              lang: lang)
-        end
-        new_arr
-      end
-
-      def symbol_to_text(symbol, lang:, options:, intent: false)
-        case lang
-        when :asciimath
-          symbol.to_asciimath(options: options)
-        when :latex
-          symbol.to_latex(options: options)
-        when :mathml
-          symbol.to_mathml_without_math_tag(intent,
-                                            options: options).nodes.first
-        when :omml
-          symbol.to_omml_without_math_tag(true, options: options)
-        when :unicodemath
-          symbol.to_unicodemath(options: options)
-        end
-      end
-
       def primes_constants
         primes = {}
         primes
@@ -417,18 +342,6 @@ module Plurimath
         field.input(:unicodemath)&.flatten&.find do |input|
           input.match?(/&#x.+;/)
         end
-      end
-
-      def notations_to_mask(notations)
-        mask = notations.split.map do |notation|
-          MASK_CLASSES.key(notation)
-        end
-        mask.inject(*:+) ^ 15
-      end
-
-      def math_display_text_objects(object)
-        class_names = ["plus", "minus", "circ", "equal", "symbol"].freeze
-        class_names.include?(object.class_name)
       end
 
       def latex_table_curly_paren(string)
